select_matches method has been modified to make sure the reference clip is always included and all clips previously validated by the user are also included.
Changed "if ticket["tuning_update"]"" to "if "tuning_update" in ticket" to fix an error.
Add all matches scored by the user to the final report, regardless of how the algorithm thinks that clip should be scored.
Note the user score in a new column of the final report, and use "Algorithm match" to denote matching clips that were
not scored explicitly by the user.
Multiple comments added.